Update hours column of time_of_day to REAL not INTEGER

+++ data_files/example_dbs/materials.sql
@@ -1347,7 +1347,7 @@ CREATE TABLE time_of_day
     sequence INTEGER UNIQUE,
     tod      TEXT
         PRIMARY KEY,
-    hours    INTEGER NOT NULL DEFAULT 1,
+    hours    REAL NOT NULL DEFAULT 1,
     notes    TEXT
 );
 INSERT INTO "time_of_day" (sequence, tod, hours) VALUES(1,'morning',6);
